Refactoring services

+++ Forms.Application/Services/LikedTemplateService.cs
@@ -1,5 +1,6 @@
 using Forms.Application.DTOs;
 using Forms.Application.Interfaces.IServices;
+using Forms.Application.Mapping;
 using Forms.Core.Interfaces.IRepositories;
 using Forms.Core.Models;
 
@@ -10,33 +11,31 @@ public class LikedTemplateService(ILikedTemplateRepository repository, ITemplate
     public async Task<List<Template>> GetLikedTemplates(uint? userId)
     {
         if (userId == null) {throw new ArgumentException("Incorrect user id");}
-        return await repository.GetLikedTemplatesByUserId(userId.Value);
+        var templates = await repository.GetLikedTemplatesByUserId(userId.Value);
+        if (templates == null) {throw new ArgumentException("Templates not found");}
+        return templates;
     }
 
     public async Task AddLikedTemplate(LikedTemplateDto likedTemplateDto)
     {
-        var checkLikedTemplate = await repository.GetLikedTemplate(likedTemplateDto.UserId, likedTemplateDto.TemplateId);
+        if(likedTemplateDto.UserId == null) throw new ArgumentException("Incorrect user id");
+        if(likedTemplateDto.TemplateId == null) throw new ArgumentException("Incorrect template id");
+        var checkLikedTemplate = await repository.GetLikedTemplate(likedTemplateDto.UserId.Value, likedTemplateDto.TemplateId.Value);
 
         if (checkLikedTemplate is not null) {throw new InvalidOperationException("Template already liked");}
-        var likedTemplate = new LikedTemplate
-        {
-            UserId = likedTemplateDto.UserId,
-            TemplateId = likedTemplateDto.TemplateId
-        };
+        var likedTemplate = LikedTemplateMapping.AddLikedTemplate(likedTemplateDto);
         await repository.AddLikedTemplate(likedTemplate);
         await templateService.IncreaseLikeNumber(likedTemplateDto.TemplateId);
     }
 
     public async Task RemoveLikedTemplate(LikedTemplateDto likedTemplateDto)
     {
-        var checkLikedTemplate = await repository.GetLikedTemplate(likedTemplateDto.UserId, likedTemplateDto.TemplateId);
+        if(likedTemplateDto.UserId == null) throw new ArgumentException("Incorrect user id");
+        if(likedTemplateDto.TemplateId == null) throw new ArgumentException("Incorrect template id");
+        var checkLikedTemplate = await repository.GetLikedTemplate(likedTemplateDto.UserId.Value, likedTemplateDto.TemplateId.Value);
 
         if (checkLikedTemplate is null) {throw new InvalidOperationException("Template not already liked");}
-        var likedTemplate = new LikedTemplate
-        {
-            UserId = likedTemplateDto.UserId,
-            TemplateId = likedTemplateDto.TemplateId
-        };
+        var likedTemplate = LikedTemplateMapping.AddLikedTemplate(likedTemplateDto);
         await repository.RemoveLikedTemplate(likedTemplate);
         await templateService.DecreaseLikeNumber(likedTemplateDto.TemplateId);
     }
